backend/requirements-incomplete.txt
django==3.2
graphene-django
django-filter
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
pyjwt
dj-database-url
django-ratelimit
colorlog
dj-static
django-extensions
django-silk
cachetools
pandas
feedparser
gensim
autopep8
pylint-django
pytest-django
mixer
pytest-cov
flake8
werkzeug
scikit-learn==0.21.2
deeppavlov==0.17.3
transformers==4.16.2
sentry-sdk==1.5.8
celery==5.1.2
django-celery-beat==2.2.1
eventlet==0.30.2
gunicorn==20.0.4
psycopg2-binary
